% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_season_column}
\alias{add_season_column}
\title{Add a column for the season}
\usage{
add_season_column(data)
}
\arguments{
\item{data}{A data frame containing a `month` column.}
}
\value{
The modified data frame with `season` column added.
}
\description{
Add a column for the season
}
