% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_year_range_column}
\alias{add_year_range_column}
\title{Add a column for year ranges}
\usage{
add_year_range_column(data)
}
\arguments{
\item{data}{A data frame containing a `year` column.}
}
\value{
The modified data frame with `range_of_years` column added.
}
\description{
This function adds a column to a data frame that categorizes years into specified ranges.
}
