% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analyze.R
\name{calculate_mean_results}
\alias{calculate_mean_results}
\title{Calculate mean results for the data}
\usage{
calculate_mean_results(data)
}
\arguments{
\item{data}{A data frame containing the data.}
}
\value{
A data frame with mean results.
}
\description{
Calculate mean results for the data
}
