% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analyze.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate statistical parameters (mean, median, standard deviation)}
\usage{
calculate_statistics(df, parameter)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{parameter}{The column name to group by (unquoted).}
}
\value{
A data frame with calculated statistics.
}
\description{
This function calculates statistical parameters (mean, median, standard deviation) for each group defined by the given parameter and `type`.
}
\examples{
\dontrun{
data <- data.frame(type = rep(c("A", "B"), each = 5), value = rnorm(10), year = rep(2001:2005, 2))
calculate_statistics(data, year)
}
}
