% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{combine_data_frames}
\alias{combine_data_frames}
\title{Combine a list of data frames into one}
\usage{
combine_data_frames(processed_df)
}
\arguments{
\item{processed_df}{A list of processed data frames.}
}
\value{
A combined data frame.
}
\description{
Combine a list of data frames into one
}
