% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analyze.R
\name{count_tau_kendall}
\alias{count_tau_kendall}
\title{Calculate Kendall's Tau coefficient for two variables}
\usage{
count_tau_kendall(data)
}
\arguments{
\item{data}{A data frame containing `year` and `value` columns.}
}
\value{
The Kendall's Tau coefficient.
}
\description{
Calculate Kendall's Tau coefficient for two variables
}
