% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{create_stat_plot}
\alias{create_stat_plot}
\title{Create a statistical plot}
\usage{
create_stat_plot(statistics_df, stat_col, stat_name)
}
\arguments{
\item{statistics_df}{A data frame with statistical data.}

\item{stat_col}{The column for the statistic.}

\item{stat_name}{The name of the statistic.}
}
\value{
A ggplot object showing the statistical plot.
}
\description{
Create a statistical plot
}
