% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_mean_plot}
\alias{draw_mean_plot}
\title{Draw a plot of average values within types for different time ranges}
\usage{
draw_mean_plot(data, mean_results)
}
\arguments{
\item{data}{A data frame containing the original data.}

\item{mean_results}{A data frame containing the mean results.}
}
\value{
A ggplot object showing the mean values plot.
}
\description{
Draw a plot of average values within types for different time ranges
}
