% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_scatter_with_trend_plot}
\alias{draw_scatter_with_trend_plot}
\title{Draw a scatter plot with a trend line for values}
\usage{
draw_scatter_with_trend_plot(data)
}
\arguments{
\item{data}{A data frame containing the data.}
}
\value{
A ggplot object showing the scatter plot with a trend line.
}
\description{
Draw a scatter plot with a trend line for values
}
