% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_stat_plot}
\alias{draw_stat_plot}
\title{Create a statistical plot}
\usage{
draw_stat_plot(statistics_df, stat_col, stat_name)
}
\arguments{
\item{statistics_df}{A data frame with statistical data.}

\item{stat_col}{The column name for the statistic to plot (unquoted).}

\item{stat_name}{The name of the statistic to display in the plot title.}
}
\value{
A ggplot object showing the statistical plot.
}
\description{
This function creates a line plot showing the given statistical parameter across different time ranges and types.
}
\examples{
\dontrun{
statistics_df <- data.frame(range_of_years = rep(c("2004-2006", "2007-2009"), each = 2),
                            type = rep(c("A", "B"), 2), mean_value = runif(4))
draw_stat_plot(statistics_df, mean_value, "Średnia")
}
}
