% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_tao}
\alias{draw_tao}
\title{Draw a Tau plot}
\usage{
draw_tao(data)
}
\arguments{
\item{data}{A data frame containing the data.}
}
\value{
A ggplot object showing the Tau plot.
}
\description{
Draw a Tau plot
}
