% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_time_series_plot}
\alias{draw_time_series_plot}
\title{Draw a time series plot of values for different types}
\usage{
draw_time_series_plot(data)
}
\arguments{
\item{data}{A data frame containing `year`, `value`, and `type` columns.}
}
\value{
A ggplot object showing the time series plot.
}
\description{
This function creates a time series plot showing the values over the years for different types.
}
\examples{
\dontrun{
data <- data.frame(year = rep(2001:2005, each = 2), value = rnorm(10), type = rep(c("A", "B"), 5))
draw_time_series_plot(data)
}
}
