% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_df}
\alias{process_df}
\title{Process a list of data frames}
\usage{
process_df(data_list)
}
\arguments{
\item{data_list}{A list of data frames.}
}
\value{
A list of processed data frames.
}
\description{
Process a list of data frames
}
