% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{read_all_files}
\alias{read_all_files}
\title{Load all CSV files from the specified folder}
\usage{
read_all_files(folder_path)
}
\arguments{
\item{folder_path}{Character. The path to the folder containing CSV files.}
}
\value{
A list containing two elements: `data_list` (a list of data frames loaded from CSV files) and `files` (a character vector of filenames).
}
\description{
This function loads all CSV files from the specified folder into a list of data frames.
}
