% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{split_date}
\alias{split_date}
\title{Split date into month and year}
\usage{
split_date(data)
}
\arguments{
\item{data}{A data frame containing a date column.}
}
\value{
The modified data frame with `year` and `month` columns added.
}
\description{
Split date into month and year
}
