% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analyze.R
\name{calculate_mean_results}
\alias{calculate_mean_results}
\title{Calculate mean results for the data}
\usage{
calculate_mean_results(data)
}
\arguments{
\item{data}{A data frame containing the data.}
}
\value{
A data frame with mean results.
}
\description{
This function calculates the mean values of the data grouped by `range_of_years` and `type`.
}
\examples{
\dontrun{
data <- data.frame(range_of_years = rep(c("2004-2006", "2007-2009"), each = 5), type = rep(c("A", "B"), each = 5), value = rnorm(10))
calculate_mean_results(data)
}
}
