% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_mean_plot}
\alias{draw_mean_plot}
\title{Draw a plot of average values within types for different time ranges}
\usage{
draw_mean_plot(data, mean_results)
}
\arguments{
\item{data}{A data frame containing the original data.}

\item{mean_results}{A data frame containing the mean results.}
}
\value{
A ggplot object showing the mean values plot.
}
\description{
This function creates a line plot showing the average values within types for different time ranges.
}
\examples{
\dontrun{
data <- data.frame(range_of_years = rep(c("2004-2006", "2007-2009"), each = 5),
                   value = rnorm(10), type = rep(c("A", "B"), 5))
mean_results <- data.frame(range_of_years = rep(c("2004-2006", "2007-2009"), each = 2),
                           mean_value = runif(4), type = rep(c("A", "B"), 2))
draw_mean_plot(data, mean_results)
}

}
