% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_scatter_plot}
\alias{draw_scatter_plot}
\title{Draw a scatter plot of values for different types across time ranges}
\usage{
draw_scatter_plot(data)
}
\arguments{
\item{data}{A data frame containing the data.}
}
\value{
A ggplot object showing the scatter plot.
}
\description{
Draw a scatter plot of values for different types across time ranges
}
