% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_scatter_with_trend_plot}
\alias{draw_scatter_with_trend_plot}
\title{Draw a scatter plot with a trend line for values}
\usage{
draw_scatter_with_trend_plot(data)
}
\arguments{
\item{data}{A data frame containing `year`, `value`, and `type` columns.}
}
\value{
A ggplot object showing the scatter plot with a trend line.
}
\description{
This function creates a scatter plot with a linear trend line showing the values over the years for different types.
}
\examples{
\dontrun{
data <- data.frame(year = rep(2001:2005, each = 2), value = rnorm(10), type = rep(c("A", "B"), 5))
draw_scatter_with_trend_plot(data)
}
}
