% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_season_plot}
\alias{draw_season_plot}
\title{Draw a seasonal plot}
\usage{
draw_season_plot(data)
}
\arguments{
\item{data}{A data frame containing the data.}
}
\value{
A ggplot object showing the seasonal plot.
}
\description{
Draw a seasonal plot
}
