% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{draw_value_distribution_plot}
\alias{draw_value_distribution_plot}
\title{Draw a distribution plot of values for different types across time ranges}
\usage{
draw_value_distribution_plot(data)
}
\arguments{
\item{data}{A data frame containing the data.}
}
\value{
A ggplot object showing the value distribution plot.
}
\description{
Draw a distribution plot of values for different types across time ranges
}
