% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_df}
\alias{process_df}
\title{Process a list of data frames}
\usage{
process_df(data_list)
}
\arguments{
\item{data_list}{A list of data frames.}
}
\value{
A combined data frame.
}
\description{
This function processes a list of data frames by extracting type information, splitting date into year and month, adding year range and season columns, and converting values to numeric.
}
