% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{read_all_files}
\alias{read_all_files}
\title{Load all CSV files from the specified folder}
\usage{
read_all_files(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder containing CSV files.}
}
\value{
A list containing two lists: `data_list` with loaded data and `files` with filenames.
}
\description{
Load all CSV files from the specified folder
}
